Rewrite the incremental LR engine.
We now implement the same incremental engine as the OCaml table backend
of Menhir, exposing the same set of features. However, because some
functions could leave the parser in an invalid state if used
incorrectly, this interface is still considered unstable and should not
be used directly by the programmer. It's used only to implement the
stable, safe monolithic interface.

SemAct is now an alias for the function pointer alone, instead of being
an option. It makes more sense (an action is the code itself) and allows
to use the pointer type alone in the new engine with an handy alias.